pyproject.toml
[project]
name = "kafka-project"
version = "0.1.0"
description = "Real-time data pipeline with Kafka, ML analysis, and Streamlit dashboard"
requires-python = ">=3.9"
dependencies = [
    "kafka-python>=2.0.2",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.0",
]

